album fertig

+++ album.php
@@ -2,11 +2,17 @@
 
 <?php
 $uploadDir = "Uploads/";
+$thumbnailDir = "thumbnails/";
+
+$thumbnailPath = $uploadDir . $thumbnailDir;
 
 if (!file_exists($uploadDir)) {
     mkdir($uploadDir);
 }
 
+if (!file_exists($thumbnailPath)) {
+    mkdir($thumbnailPath, 0777, true);
+}
 //echo "<pre>";print_r($_FILES);"</pre";
 
 //Erlaubte Dateiformate
@@ -24,22 +30,64 @@ if (
 
     //Checkt ob die gespeicherte $ext im $allowed array drin is, wenn mehr Formate erlaubt --> allowed array erweitern s. oben
     if (in_array($ext, $allowed)) {
-        if (
-            move_uploaded_file(
-                $_FILES["file"]["tmp_name"],
-                $uploadDir . $_FILES["file"]["name"]
-            )
-        ) {
-            echo 'Datei erfolgreich hochgeladen';
+        if (!file_exists($uploadDir . $filename)) {
+            if (
+                move_uploaded_file(
+                    $_FILES["file"]["tmp_name"],
+                    $uploadDir . $_FILES["file"]["name"]
+                )
+            ) {
+                $filepath = $uploadDir . $filename;
+                if (createThumbnail($filename, $filepath, $ext, $thumbnailPath)) {
+                    echo 'Thumbnail erfolgreich erstellt <br>';
+                } else {
+                    echo 'Fehler beim Erstellen des Thumbnail <br>';
+                }
+                echo 'Datei erfolgreich hochgeladen';
+            } else {
+                echo 'Fehler beim Hochladen \r\n';
+            }
         } else {
-            echo 'Fehler beim Hochladen';
+            echo 'Fehler: Die Datei "' . $filename . '" gibt es bereits.';
         }
     } else {
-        echo 'Fehler: Bitte nur folgende Bildformate hochladen: '; 
+        echo 'Fehler: Bitte nur folgende Bildformate hochladen: ';
         //gibt alle erlaubten formate aus
-        for ($i=0; $i < count($allowed); $i++) { 
-            echo '.' .$allowed[$i] .', ';
-        };
+        for ($i = 0; $i < count($allowed); $i++) {
+            echo '.' . $allowed[$i] . ', ';
+        }
+        ;
+    }
+}
+function createThumbnail($filename, $filepath, $ext, $thumbnailPath)
+{
+    // Get new sizes
+    list($width, $height) = getimagesize($filepath);
+    $newwidth = 200;
+    $newheight = 200;
+
+    // Load
+    $thumb = imagecreatetruecolor($newwidth, $newheight);
+
+    switch ($ext) {
+        case 'png':
+            $source = imagecreatefrompng($filepath);
+            break;
+
+        case 'jpeg' || 'jpg':
+            $source = imagecreatefromjpeg($filepath);
+            break;
+    }
+
+    // Resize
+    imagecopyresized($thumb, $source, 0, 0, 0, 0, $newwidth, $newheight, $width, $height);
+
+    // Output
+    switch ($ext) {
+        case 'png':
+            return imagepng($thumb, $thumbnailPath . 'thumb_' . $filename);
+        case 'jpeg' || 'jpg':
+            return imagejpeg($thumb, $thumbnailPath . 'thumb_' . $filename);
     }
 }
 ?>
@@ -98,215 +146,55 @@ if (
                 <img src="" alt="">
 
                 <div class="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
+                    <?php
+                    if (file_exists($thumbnailPath)) {
+                        $files = scandir($thumbnailPath);
+                        /*echo "<pre>";
+                        print_r($files);
+                        "</pre>";*/
+                        if (count($files) == 2) {
+                            echo "Keine Files vorhanden ...";
+                            //                echo "<li>Keine Files vorhanden ...</li>";
+                            //                echo "<p class='red' <li>Keine Files vorhanden ...</li>";
+                        }
+                        for ($i = 2; isset($files[$i]); $i++) {
+                            if ($files[$i] != "." && $files[$i] != "..") {
+                    ?>
+
                     <div class="col">
                         <div class="card shadow-sm">
-                            <img class="bd-placeholder-img card-img-top img-thumbnail" src="http://www.w3.org/2000/svg"
-                                role="img" aria-label="Placeholder: Thumbnail" preserveAspectRatio="xMidYMid slice"
+                            <img class="bd-placeholder-img card-img-top img-thumbnail"
+                                src="<?php echo $thumbnailPath . $files[$i]; ?>" role="img"
+                                aria-label="Placeholder: Thumbnail" preserveAspectRatio="xMidYMid slice"
                                 focusable="false">
-                            <title>Placeholder</title>
+                            <title>
+                                <?php echo $files[$i]; ?>
+                            </title>
                             <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                dy=".3em">Thumbnail</text>
-
-
+                                dy=".3em">
+                                <?php echo $files[$i]; ?>
+                            </text>
                             <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
+                                <p class="card-text">
+
+                                </p>
                                 <div class="d-flex justify-content-between align-items-center">
                                     <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Download</button>
+                                        <a href="<?php /*link zum Blogartikel*/?>"
+                                            class="btn btn-sm btn-outline-secondary" target=blank>Newsbeitrag lesen</a>
+                                        <a href="<?php echo $thumbnailPath . $files[$i]; ?>"
+                                            class="btn btn-sm btn-outline-secondary" target=blank>Anschauen</a>
+                                        <a href="<?php echo $thumbnailPath . $files[$i]; ?>"
+                                            class=" btn btn-sm btn-outline-secondary" download>Download</a>
                                     </div>
                                     <small class="text-muted">9 mins</small>
                                 </div>
                             </div>
                         </div>
                     </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Test</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col">
-                        <div class="card shadow-sm">
-                            <svg class="bd-placeholder-img card-img-top" width="100%" height="225"
-                                xmlns="http://www.w3.org/2000/svg" role="img" aria-label="Placeholder: Thumbnail"
-                                preserveAspectRatio="xMidYMid slice" focusable="false">
-                                <title>Placeholder</title>
-                                <rect width="100%" height="100%" fill="#55595c" /><text x="50%" y="50%" fill="#eceeef"
-                                    dy=".3em">Thumbnail</text>
-                            </svg>
-
-                            <div class="card-body">
-                                <p class="card-text">This is a wider card with supporting text below as a natural
-                                    lead-in to additional content. This content is a little bit longer.</p>
-                                <div class="d-flex justify-content-between align-items-center">
-                                    <div class="btn-group">
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-                                        <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-                                    </div>
-                                    <small class="text-muted">9 mins</small>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
+                    <?php }
+                        }
+                    } ?>
                 </div>
             </div>
         </div>
